feat(tests): migrate Unit.Tests → Web.Tests and remove Unit.Tests project
- Add Handle_Success_DoesNotCallInvalidateById to CreateBlogPostHandlerTests
- Add HandleGetById_CacheServiceThrows_ReturnsFailResult to EditBlogPostHandlerTests
- Remove InternalsVisibleTo(Unit.Tests), add Web.Tests/Bunit/Integration to AssemblyInfo.cs
- Remove Unit.Tests from solution file and squad-release.yml workflow
- Delete tests/Unit.Tests/ project entirely

Closes #151
