+++ MyBlog.slnx
@@ -9,7 +9,6 @@
     <Project Path="tests/AppHost.Tests/AppHost.Tests.csproj" />
     <Project Path="tests/Architecture.Tests/Architecture.Tests.csproj" />
     <Project Path="tests/Domain.Tests/Domain.Tests.csproj" />
-    <Project Path="tests/Unit.Tests/Unit.Tests.csproj" />
     <Project Path="tests/Web.Tests/Web.Tests.csproj" />
     <Project Path="tests/Web.Tests.Bunit/Web.Tests.Bunit.csproj" />
     <Project Path="tests/Web.Tests.Integration/Web.Tests.Integration.csproj" />
