--- tests/Unit.Tests/GlobalUsings.cs
@@ -1,18 +1,0 @@
-//=======================================================
-//Copyright (c) 2026. All rights reserved.
-//File Name :     GlobalUsings.cs
-//Company :       mpaulosky
-//Author :        Matthew Paulosky
-//Solution Name : MyBlog
-//Project Name :  Unit.Tests
-//=======================================================
-
-global using FluentAssertions;
-global using Microsoft.EntityFrameworkCore;
-global using MyBlog.Domain.Abstractions;
-global using MyBlog.Domain.Entities;
-global using MyBlog.Domain.Interfaces;
-global using MyBlog.Web.Data;
-global using MyBlog.Web.Infrastructure.Caching;
-global using NSubstitute;
-global using NSubstitute.ExceptionExtensions;
