--- tests/Unit.Tests/Handlers/CreateBlogPostHandlerTests.cs
@@ -1,73 +1,0 @@
-//=======================================================
-//Copyright (c) 2026. All rights reserved.
-//File Name :     CreateBlogPostHandlerTests.cs
-//Company :       mpaulosky
-//Author :        Matthew Paulosky
-//Solution Name : MyBlog
-//Project Name :  Unit.Tests
-//=======================================================
-
-// TDD-red: these tests will not compile until #110 refactors CreateBlogPostHandler
-// to accept IBlogPostCacheService instead of IMemoryCache + IDistributedCache.
-
-using MyBlog.Web.Features.BlogPosts.Create;
-
-namespace Unit.Handlers;
-
-public class CreateBlogPostHandlerTests
-{
-	private readonly IBlogPostRepository _repo = Substitute.For<IBlogPostRepository>();
-	private readonly IBlogPostCacheService _cache = Substitute.For<IBlogPostCacheService>();
-	private readonly CreateBlogPostHandler _handler;
-
-	public CreateBlogPostHandlerTests()
-	{
-		_handler = new CreateBlogPostHandler(_repo, _cache);
-	}
-
-	[Fact]
-	public async Task Handle_Success_CreatesPostAndInvalidatesAll()
-	{
-		// Arrange
-		var command = new CreateBlogPostCommand("Title", "Content", "Author");
-
-		// Act
-		var result = await _handler.Handle(command, CancellationToken.None);
-
-		// Assert
-		result.Success.Should().BeTrue();
-		result.Value.Should().NotBeEmpty();
-		await _repo.Received(1).AddAsync(Arg.Any<BlogPost>(), Arg.Any<CancellationToken>());
-		await _cache.Received(1).InvalidateAllAsync(Arg.Any<CancellationToken>());
-	}
-
-	[Fact]
-	public async Task Handle_Success_DoesNotCallInvalidateById()
-	{
-		// Arrange — create should only bust the "all" list, not a specific post key
-		var command = new CreateBlogPostCommand("Title", "Content", "Author");
-
-		// Act
-		await _handler.Handle(command, CancellationToken.None);
-
-		// Assert
-		await _cache.DidNotReceive().InvalidateByIdAsync(Arg.Any<Guid>(), Arg.Any<CancellationToken>());
-	}
-
-	[Fact]
-	public async Task Handle_RepoThrows_ReturnsFailResult()
-	{
-		// Arrange
-		var command = new CreateBlogPostCommand("Title", "Content", "Author");
-		_repo.AddAsync(Arg.Any<BlogPost>(), Arg.Any<CancellationToken>())
-			.ThrowsAsync(new InvalidOperationException("insert failed"));
-
-		// Act
-		var result = await _handler.Handle(command, CancellationToken.None);
-
-		// Assert
-		result.Failure.Should().BeTrue();
-		result.Error.Should().Contain("insert failed");
-		await _cache.DidNotReceive().InvalidateAllAsync(Arg.Any<CancellationToken>());
-	}
-}
