--- tests/Unit.Tests/Unit.Tests.csproj
@@ -1,34 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<TargetFramework>net10.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-		<IsPackable>false</IsPackable>
-		<IsTestProject>true</IsTestProject>
-		<RootNamespace>MyBlog.Unit.Tests</RootNamespace>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="coverlet.collector"/>
-		<PackageReference Include="coverlet.msbuild">
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-			<PrivateAssets>all</PrivateAssets>
-		</PackageReference>
-		<PackageReference Include="FluentAssertions"/>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" />
-		<PackageReference Include="NSubstitute"/>
-		<PackageReference Include="xunit"/>
-		<PackageReference Include="xunit.runner.visualstudio"/>
-	</ItemGroup>
-
-	<ItemGroup>
-		<Using Include="Xunit"/>
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\..\src\Domain\Domain.csproj"/>
-		<ProjectReference Include="..\..\src\Web\Web.csproj"/>
-	</ItemGroup>
-
-</Project>
